refactor: share env var guard for tests
