+++ crates/luban_backend/src/lib.rs
@@ -1,6 +1,8 @@
 mod env;
 mod services;
 mod sqlite_store;
+#[cfg(test)]
+mod test_support;
 mod time;
 
 pub use services::GitWorkspaceService;
